Allow configuration loading from Rails config/ directory
In order to keep the plugin updatable (i.e. as a submodule in
my Redmine installation), it is preferable to put the configuration
file in the Rails config/ directory instead of the plugin directory.
With the previous configuration loading, one would have to fork the
plugin in order to keep "updatability" through a submodule.